Fixed reload, improved styling and working on comments from feecback
