+++ index.html
@@ -8,44 +8,44 @@
     <!-- <script src="./node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script> -->
     <script type="module" src="./src/js/index.mjs"></script>
   </head>
-  <body class="background-primary">
-    <main class="container vh-100 d-flex justify-content-center">
-      <div class="row h-100 background-secondary main-width">
-        <div class="col-12 d-flex justify-content-center align-items-center mt-5">
-          <img src="./src/img/png/logo-color.png" height="280" width="280" class="rounded-circle size-md-2" />
+  <body>
+    <main class="container-fluid vh-100 d-flex justify-content-center background-primary">
+      <div class="row h-100 main-width">
+        <div class="col-12 text-center pt-3 custom-fixed-height">
+          <img src="./src/img/png/logo-color.png" height="280" width="280" alt="Website Logo" class="rounded-circle" />
         </div>
-
-        <div class="col-12 w-100">
+        <div class="col-12 custom-fixed-lower">
           <div class="custom-card text-center border-0 bg-transparent">
-            <div class="card-body login-card-border mt-4">
+            <div class="card-body login-card-border mt-3">
               <div class="row">
                 <div class="col">
                   <div class="collapse show multi-collapse" id="multiCollapseExample1">
                     <!-- Log in form -->
-                    <h5 class="card-title">Log in</h5>
+                    <h5 class="card-title custom-title-font fs-3 text-primary">Log in!</h5>
+                    <span id="indexErrorMessages"></span>
                     <form class="row g-3 needs-validation" id="loginForm" autocomplete="on" type="submit" action="./post&feed/post&feed.html" method="post" novalidate>
                       <div class="mb-3">
-                        <input type="email" class="form-control" autocomplete="email" id="logInEmail" placeholder="name@example.noroff" required />
+                        <input type="email" class="form-control custom-text-font" autocomplete="email" id="logInEmail" placeholder="name@example.noroff" required />
 
-                        <div class="invalid-feedback">Please enter valid email</div>
+                        <div class="invalid-feedback text-black fs-6">Please enter valid email</div>
                       </div>
                       <div class="mb-3">
-                        <input type="password" autocomplete="current-password" class="form-control" id="logInPassword" placeholder="Password" required />
-                        <div class="invalid-feedback">
+                        <input type="password" autocomplete="current-password" class="form-control custom-text-font" id="logInPassword" placeholder="Password" required />
+                        <div class="invalid-feedback text-black fs-6">
                           Please enter correct password
-                          <button class="btn text-center w-100">Forgot Password?</button>
+                          <button class="btn text-center w-100 custom-text-font text-danger fs-5 fw-bold">Forgot Password?</button>
                         </div>
                       </div>
 
-                      <div class="col-12 my-4">
-                        <button class="btn bg-primary" type="submit">Log in</button>
+                      <div class="col-12 mb-4">
+                        <button class="btn bg-primary custom-title-font text-white" type="submit">Log in</button>
                       </div>
                     </form>
                     <div class="row">
                       <div class="col-5">
                         <hr />
                       </div>
-                      <div class="col-2">Or</div>
+                      <div class="col-2 custom-title-font fs-6">Or</div>
                       <div class="col-5">
                         <hr />
                       </div>
@@ -68,43 +68,42 @@
                     Log in
                   </button>
                   <h5 class="card-title">Create account</h5>
+                  <span id="indexErrorMessages2"></span>
                   <form class="row g-3 needs-validation" id="signUpForm" autocomplete="on" method="post" novalidate>
                     <div class="mb-3">
-                      <div id="debugOutput"></div>
                       <input type="text" class="form-control" autocomplete="name" id="signUpUserName" placeholder="User name" required />
-                      <div class="invalid-feedback" id="usernameFeedback">Please enter a user name</div>
+                      <div class="invalid-feedback text-black fs-6" id="usernameFeedback">Please enter a valid user name</div>
                     </div>
                     <div class="mb-3">
                       <input type="email" autocomplete="email" class="form-control" id="signUpEmail" placeholder="noroff.no / stud.noroff.no" required />
-                      <div class="invalid-feedback" id="invalidEmailFeedback"></div>
+                      <div class="invalid-feedback text-black fs-6" id="invalidEmailFeedback"></div>
                     </div>
                     <div class="mb-3">
                       <input type="password" autocomplete="new-password" class="form-control" id="signUpPassword" minlength="8" placeholder="New password" required />
-                      <div class="invalid-feedback" id="">Minimum 8 character password</div>
+                      <div class="invalid-feedback text-black fs-6" id="">Minimum 8 character password</div>
                     </div>
                     <div class="mb-3">
                       <input type="password" autocomplete="new-password" class="form-control" id="signUpConfirmPassword" placeholder="Confirm password" required />
-                      <div class="invalid-feedback" id="passwordConfirmFeedback">Passwords does not match</div>
+                      <div class="invalid-feedback text-black fs-6" id="passwordConfirmFeedback">Passwords does not match</div>
                     </div>
 
                     <div class="col-12 my-4">
-                      <button class="btn btn-primary" id="sign-up-btn" type="submit">Real Sign up!</button>
+                      <button class="btn btn-primary" id="sign-up-btn" type="submit">Sign up!</button>
                     </div>
                   </form>
                 </div>
+                <button
+                  class="btn btn-primary my-4"
+                  id="collapsable-btn"
+                  type="button"
+                  data-bs-toggle="collapse"
+                  data-bs-target=".multi-collapse"
+                  aria-expanded="false"
+                  aria-controls="multiCollapseExample1 multiCollapseExample2"
+                >
+                  Sign Up
+                </button>
               </div>
-
-              <button
-                class="btn btn-primary my-4"
-                id="collapsable-btn"
-                type="button"
-                data-bs-toggle="collapse"
-                data-bs-target=".multi-collapse"
-                aria-expanded="false"
-                aria-controls="multiCollapseExample1 multiCollapseExample2"
-              >
-                Sign Up
-              </button>
             </div>
           </div>
         </div>
