+++ src/js/index.mjs
@@ -1,10 +1,7 @@
-import {
-  setLoginFormDataCallback,
-  setSignUpFormDataCallback,
-} from "./utilities/forms.mjs";
+import { setLoginFormDataCallback, setSignUpFormDataCallback } from "./utilities/forms.mjs";
 import { collapsableForm } from "./utilities/formcollapse.mjs";
-import { wrappedCreateAccount } from "./utilities/createAccount.mjs";
-import { wrappedLogInUser } from "./utilities/logIn.mjs";
+import { wrappedCreateAccount } from "./utilities/create-account.mjs";
+import { wrappedLogInUser } from "./utilities/log-in.mjs";
 
 setLoginFormDataCallback();
 setSignUpFormDataCallback();
