+++ src/js/utilities/sign-out.mjs
