+++ src/scss/main.scss
@@ -1,5 +1,5 @@
-@import "./partials/editedVariables.scss";
+@import "./partials/editedvariables.scss";
 @import "./partials/index";
-@import "./partials/loginCard";
-@import "./partials/post&feed";
-@import "./partials/SoMeCard";
+@import "./partials/logincard";
+@import "./partials/postandfeed";
+@import "./partials/somecard";
