+++ src/scss/partials/_index.scss
@@ -11,5 +11,17 @@
 }
 
 .custom-fixed-height {
-  height: 600px;
+  height: 330px;
 }
+
+.custom-fixed-lower {
+  height: 550px;
+}
+
+.logo-front-page {
+  height: 50%;
+}
+
+// .custom-max-height {
+//   max-height: 1000px !important;
+// }
